Return long instead of int when we are counting countries by name on Russian.
Should be in 7434210c5fd10af188a0b01b98c172c219283c57 commit.

No functional changes.

+++ src/main/java/ru/mystamps/web/dao/CountryDao.java
@@ -26,7 +26,7 @@ import ru.mystamps.web.service.dto.SelectEntityDto;
 
 public interface CountryDao extends PagingAndSortingRepository<Country, Integer> {
 	long countByName(String name);
-	int countByNameRu(String name);
+	long countByNameRu(String name);
 	
 	@Query(
 		"SELECT NEW ru.mystamps.web.service.dto.SelectEntityDto("
